implemented native clients
