+++ lib/KafkaFactory.js
@@ -1,6 +1,7 @@
 "use strict";
 
 const KafkaClient = require("./KafkaClient.js");
+const NativeKafkaClient = require("./NativeKafkaClient.js");
 
 class KafkaFactory {
 
@@ -19,6 +20,11 @@ class KafkaFactory {
     }
 
     getKafkaClient(topic){
+
+        if(this.config.noptions && typeof this.config.noptions === "object"){
+            return new NativeKafkaClient(topic, this.config);
+        }
+
        return new KafkaClient(topic, this.config);
     }
 }
