+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kafka-streams",
-  "version": "2.3.0",
+  "version": "3.0.0",
   "description": "kafka-streams for nodejs",
   "main": "index.js",
   "scripts": {
